requirements.txt
# Langchain Core
langchain 
langchain-core

# OpenAI Integration
langchain-openai 
openai

# Anthropic Integration
langchain-Anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face integration
langchain-huggingface
transformers
sentence-transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Leawrning Utilities
numpy
scikit-learn